+++ CHANGELOG.md
@@ -46,6 +46,9 @@ CHANGELOG
   
 - Fix logout with file backend when state is deleted
   [#4218](https://github.com/pulumi/pulumi/pull/4218)
+  
+- Fix specific flags for `pulumi stack` being global
+  [#4218](https://github.com/pulumi/pulumi/pull/4218)
 
 ## 1.14.0 (2020-04-01)
 - Fix error related to side-by-side versions of `@pulumi/pulumi`.